requirements.txt
ttkbootstrap
beautifulsoup4
requests
pathvalidate
PyYAML
platformdirs
pyinstaller
